[catalog] refactor package name
