+++ streampark-flink/pom.xml
@@ -63,6 +63,12 @@
             <version>${jupiter.version}</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.apache.flink</groupId>
+            <artifactId>flink-table-common</artifactId>
+            <version>1.19.0</version>
+            <scope>compile</scope>
+        </dependency>
 
     </dependencies>
