+++ streampark-flink/streampark-flink-catalog-store/src/main/resources/META-INF/services/org.apache.flink.table.factories.Factory
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-org.apache.streampark.JdbcCatalogStoreFactory
+org.apache.streampark.catalog.JdbcCatalogStoreFactory
